S15.f.6: extract plugin.ts in-process tool-calling loop (6104 → 5598 LOC)
Sixth sub-slice of S15.f.

cli/src/core/agt-task-loop.ts (new, 540 LOC):
  processTaskWithTools(taskContent, deps, log) — the AGT sub-agent
  in-process tool-calling loop (25-round LLM/tool execution) that
  every offload and parent-driven sub-task hits.
  TaskLoopDeps interface (meshClient, isInterruptRequested,
  interruptReason, setInterrupt).

plugin.ts keeps a thin wrapper that builds the deps bag from the
plugin-scope singletons. Pure imports (TASK_TOOLS, routerUrl,
resolveAmidByName, sanitizeLog) bind directly in the new module.

resolveAmidByName called with canonical 2-arg signature
(name, routerUrl, opts?) inside the new module since the
plugin-scope wrapper is no longer reachable from core/.

Cumulative S15.f delta:
  pre:  7139
  f.1:  6974  (-165)
  f.2:  6890   (-84)
  f.3:  6648  (-242)
  f.4:  6488  (-160)
  f.5:  6104  (-384)
  f.6:  5598  (-506)
  Cumulative: -1541. 2598 LOC remaining to §4.2 cap of 3000.

Verification:
  - tsc --noEmit clean / lint 30 (was 29, +1 cross-module any) / build clean
  - npm test: 454 pass / 2 skipped (baseline)

Audit: docs/security-audits/2026-04-29-phase2-hotspot-plugin-cli-f6.md
