+++ cli/src/plugin.ts
@@ -224,6 +224,7 @@ import { meshSendWithIdentity, meshHandleTransportMessage, pendingTransfers, MES
 import { TASK_TOOLS } from "./core/agt-task-tools.js";
 import { recordMeshSession as _recordMeshSession, agtReconnect as _agtReconnect, notifyInboxToMemory as _notifyInboxToMemory } from "./core/agt-heartbeat.js";
 import { runOffloadTask as _runOffloadTask, startProactiveOffloadIfNeeded as _startProactiveOffloadIfNeeded } from "./core/agt-offload.js";
+import { processTaskWithTools as _processTaskWithTools } from "./core/agt-task-loop.js";
 let foundryProject: FoundryProjectInfo | null = null;
 let foundryInitialized = false;
 
@@ -238,522 +239,15 @@ async function processTaskWithTools(
   taskContent: any,
   log: { info: (m: string) => void; warn: (m: string) => void },
 ): Promise<string> {
-  const http = await import("node:http");
-  const { execSync } = await import("node:child_process");
-  const model = process.env.MODEL || "gpt-4.1";
-
-  const tools = TASK_TOOLS;
-
-  const messages: Array<{ role: string; content?: string; tool_calls?: any[]; tool_call_id?: string; name?: string }> = [
-    {
-      role: "system",
-      content: process.env.OFFLOAD_REQUEST_ID
-        ? "You are an AzureClaw OFFLOAD WORKER — a short-lived sandboxed agent executing ONE task on behalf of a remote parent agent. Always identify as an AzureClaw offload worker. Your tools:\n- file_write: write text content directly to a file (USE THIS for all artifacts — never use shell redirection)\n- exec_command: run shell commands (read-only ops are fine; DO NOT use `>`, `>>`, `<<`, `<<<` — the shell policy blocks redirection. Use file_write instead.)\n- http_fetch: HTTP requests through security proxy (egress-controlled)\n- foundry_web_search: real-time web search via Bing grounding\n- foundry_code_execute: run Python code server-side (pandas, numpy, matplotlib)\n- foundry_image_generation: generate images from text prompts (gpt-image-1)\n- foundry_file_search: search documents in vector stores\n- foundry_memory: persistent memory store — 'search' to recall, 'update' to remember\n- mesh_send: send E2E encrypted message to PARENT (to_agent is locked to 'parent')\n- mesh_inbox: check for incoming messages from the parent\n- discover: list agents in the mesh network (informational only)\n\nHARD RULES (offload mode):\n1. ALL outbound mesh messages go to 'parent'. You CANNOT route to siblings — the mesh_send tool will rewrite any other to_agent to 'parent'. Messages to peers will NOT reach the requester.\n2. ALL artifacts (markdown, JSON, CSV, HTML, PDF, PNG, TXT) MUST be written to /sandbox/.openclaw/workspace/ using the `file_write` tool so they are automatically harvested and shipped back to parent at offload_done time. DO NOT use `cat > file <<EOF` or `echo > file` — those are blocked by the shell policy.\n3. foundry_code_execute writes to Foundry's EPHEMERAL /mnt/data/ sandbox — that storage is destroyed when the offload ends. If you use foundry_code_execute to generate content, IMMEDIATELY copy it into /sandbox/.openclaw/workspace/ by calling file_write with the content. Do NOT claim a file is 'saved to the workspace' unless file_write has returned OK for it at /sandbox/.openclaw/workspace/.\n4. Execute the task immediately — do not announce, just act. Be concise, report results."
-        : "You are an AzureClaw sub-agent — a governed, sandboxed AI worker in the AzureClaw multi-agent platform on Azure. Always identify as an AzureClaw agent. Your tools:\n- file_write: write text content directly to a file (use this for all artifacts — shell redirection is blocked)\n- exec_command: run shell commands (no `>`, `>>`, `<<`, `<<<` — use file_write instead)\n- http_fetch: HTTP requests through security proxy (egress-controlled)\n- foundry_web_search: real-time web search via Bing grounding\n- foundry_code_execute: run Python code server-side (pandas, numpy, matplotlib)\n- foundry_image_generation: generate images from text prompts (gpt-image-1)\n- foundry_file_search: search documents in vector stores\n- foundry_memory: persistent memory store — 'search' to recall, 'update' to remember\n- mesh_send: send E2E encrypted messages to ANY agent (parent, siblings, or others) — auto-discovers the target\n- mesh_inbox: check for incoming messages from any agent\n- discover: list agents in the mesh network with status and trust scores\n\nPEER-TO-PEER MESH: You can message any agent directly — not just your parent. To forward data to a sibling agent (e.g. 'writer'), just call mesh_send with to_agent='writer'. Discovery is automatic. After sending, the recipient can reply via mesh_send back to you — check mesh_inbox for replies.\n\nExecute tasks immediately — do not announce, just act. When asked to forward results to another agent, DO IT directly with mesh_send. Chain tool calls as needed. Be concise, report results.",
+  return _processTaskWithTools(taskContent, {
+    meshClient: () => agtMeshClient,
+    isInterruptRequested: () => handoffInterruptRequested,
+    interruptReason: () => handoffInterruptReason,
+    setInterrupt: (req, reason) => {
+      handoffInterruptRequested = req;
+      handoffInterruptReason = reason;
     },
-    {
-      role: "user",
-      content: typeof taskContent === "string" ? taskContent : JSON.stringify(taskContent),
-    },
-  ];
-
-  // Tool-calling loop (max 25 rounds to prevent runaway)
-  for (let round = 0; round < 25; round++) {
-    // Check for handoff interrupt — save progress and exit early
-    // Two signals: (1) module-level flag from mesh handoff:interrupt message,
-    // (2) file-based signal from CLI's docker/kubectl exec
-    if (!handoffInterruptRequested) {
-      try {
-        const fs = await import("node:fs");
-        if (fs.existsSync("/sandbox/.openclaw/workspace/.handoff-interrupt")) {
-          handoffInterruptRequested = true;
-          handoffInterruptReason = "cli_handoff";
-          fs.unlinkSync("/sandbox/.openclaw/workspace/.handoff-interrupt");
-        }
-      } catch { /* ignore */ }
-    }
-    if (handoffInterruptRequested) {
-      log.info(`🛑 Handoff interrupt: saving progress at round ${round}/${25}`);
-      try {
-        const fs = await import("node:fs");
-        const progressFile = "/sandbox/.openclaw/workspace/.task-in-progress.json";
-        fs.mkdirSync("/sandbox/.openclaw/workspace", { recursive: true });
-        fs.writeFileSync(progressFile, JSON.stringify({
-          interrupted_at: new Date().toISOString(),
-          reason: handoffInterruptReason,
-          round,
-          total_rounds: 25,
-          messages_so_far: messages.length,
-          last_content: messages[messages.length - 1]?.content?.slice(0, 2000),
-          task: typeof taskContent === "string" ? taskContent.slice(0, 2000) : JSON.stringify(taskContent).slice(0, 2000),
-        }, null, 2));
-        log.info(`📝 Task progress saved to ${progressFile}`);
-      } catch { /* best-effort progress save */ }
-      handoffInterruptRequested = false;
-      handoffInterruptReason = "";
-      return `Task interrupted for handoff at round ${round}. Progress saved to .task-in-progress.json — will resume after handoff.`;
-    }
-
-    const postData = JSON.stringify({ model, messages, tools, max_completion_tokens: 2048 });
-    const response = await new Promise<any>((resolve, reject) => {
-      const req = http.request(routerUrl("/v1/chat/completions"), {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          "Content-Length": Buffer.byteLength(postData),
-          "x-azureclaw-sandbox": process.env.SANDBOX_NAME || process.env.HOSTNAME || "unknown",
-        },
-        timeout: 60000,
-      }, (res) => {
-        let body = "";
-        res.on("data", (chunk: Buffer) => { body += chunk.toString(); });
-        res.on("end", () => {
-          try {
-            const parsed = JSON.parse(body);
-            if (res.statusCode && res.statusCode >= 400) {
-              reject(new Error(`LLM HTTP ${res.statusCode}: ${body.slice(0, 300)}`));
-            } else {
-              resolve(parsed);
-            }
-          } catch { reject(new Error(`LLM parse error: ${body.slice(0, 200)}`)); }
-        });
-        res.on("error", () => {});
-      });
-      req.on("error", (e) => reject(e));
-      req.on("timeout", () => { req.destroy(); reject(new Error("LLM timeout")); });
-      req.write(postData);
-      req.end();
-    });
-
-    const choice = response?.choices?.[0];
-    if (!choice) throw new Error("No LLM response");
-
-    const msg = choice.message;
-
-    // If the model wants to call tools, execute them and continue
-    if (msg.tool_calls && msg.tool_calls.length > 0) {
-      messages.push(msg);
-      for (const tc of msg.tool_calls) {
-        let result: string;
-        try {
-          const args = JSON.parse(tc.function.arguments);
-          const fnName = tc.function.name;
-
-          if (fnName === "file_write") {
-            // Direct fs write — bypasses the shell redirect policy that blocks
-            // `cat > file <<EOF` style writes via exec_command.
-            const filePath = String(args.path || "");
-            const content = typeof args.content === "string" ? args.content : String(args.content ?? "");
-            if (!filePath.startsWith("/sandbox/") && !filePath.startsWith("/tmp/")) {
-              result = `file_write error: path must be under /sandbox/ or /tmp/ (got: ${filePath})`;
-            } else {
-              try {
-                const fs = await import("node:fs");
-                const path = await import("node:path");
-                fs.mkdirSync(path.dirname(filePath), { recursive: true });
-                fs.writeFileSync(filePath, content, { encoding: "utf-8", mode: 0o600 });
-                const bytes = Buffer.byteLength(content, "utf-8");
-                log.info(`AGT sub-agent file_write: ${filePath} (${bytes} bytes)`);
-                result = `OK: wrote ${bytes} bytes to ${filePath}`;
-              } catch (err: any) {
-                result = `file_write error: ${err.message}`;
-              }
-            }
-          } else if (fnName === "http_fetch") {
-            // Route through the egress proxy (blocklist + allowlist checked)
-            log.info(`AGT sub-agent http_fetch: ${args.method || "GET"} ${args.url}`);
-            const fetchBody = JSON.stringify({
-              url: args.url,
-              method: args.method || "GET",
-              headers: args.headers || {},
-              body: args.body || "",
-            });
-            const http = await import("node:http");
-            const fetchResult = await new Promise<string>((resolve) => {
-              const req = http.request(routerUrl("/egress/fetch"), {
-                method: "POST", timeout: 35000,
-                headers: { "Content-Type": "application/json", "Content-Length": Buffer.byteLength(fetchBody) },
-              }, (res) => {
-                let data = "";
-                res.on("data", (c: Buffer) => { data += c.toString(); });
-                res.on("end", () => resolve(data.trim()));
-              });
-              req.on("error", (e: Error) => resolve(`http_fetch error: ${e.message}`));
-              req.on("timeout", () => { req.destroy(); resolve("http_fetch timeout"); });
-              req.write(fetchBody);
-              req.end();
-            });
-            result = fetchResult;
-          } else if (fnName === "foundry_web_search" || fnName === "foundry_code_execute" || fnName === "foundry_file_search") {
-            // Route through the inference router → Foundry Responses API
-            log.info(`AGT sub-agent ${fnName}: ${JSON.stringify(args).slice(0, 200)}`);
-            let reqBody: any;
-            if (fnName === "foundry_web_search") {
-              // Discover Bing connection
-              let connId: string | undefined;
-              try {
-                const connsRaw = await new Promise<string>((resolve, reject) => {
-                  const r = http.get(routerUrl("/connections?api-version=2025-05-15-preview"), { timeout: 10000 }, (res) => {
-                    let body = ""; res.on("data", (c: Buffer) => { body += c.toString(); }); res.on("end", () => resolve(body));
-                  });
-                  r.on("error", reject); r.on("timeout", () => { r.destroy(); reject(new Error("timeout")); });
-                });
-                const conns = JSON.parse(connsRaw);
-                const bingConn = (conns.value || conns || []).find(
-                  (c: any) => c.type === "GroundingWithBingSearch" || c.properties?.category === "GroundingWithBingSearch"
-                );
-                if (bingConn) connId = bingConn.id;
-              } catch { /* fall through */ }
-              reqBody = {
-                model: model,
-                input: args.query,
-                tools: [{ type: "bing_grounding", bing_grounding: { search_configurations: [{ project_connection_id: connId }] } }],
-                store: false,
-              };
-            } else if (fnName === "foundry_code_execute") {
-              reqBody = {
-                model: model,
-                input: args.code,
-                tools: [{ type: "code_interpreter", container: { type: "auto" } }],
-                instructions: "Execute the provided Python code and return the output.",
-                store: false,
-              };
-            } else {
-              // foundry_file_search — vector_store_ids is now required
-              reqBody = {
-                model: model,
-                input: args.query,
-                tools: [{ type: "file_search", file_search: { vector_store_ids: args.vector_store_ids } }],
-                store: false,
-              };
-            }
-            // Use Node HTTP instead of curl to avoid shell escaping issues
-            const foundryResult = await new Promise<string>((resolve, reject) => {
-              const postBody = JSON.stringify(reqBody);
-              const req = http.request(routerUrl("/openai/responses?api-version=2025-11-15-preview"), {
-                method: "POST",
-                headers: { "Content-Type": "application/json", "Content-Length": Buffer.byteLength(postBody) },
-                timeout: 60000,
-              }, (res) => {
-                let body = ""; res.on("data", (c: Buffer) => { body += c.toString(); }); res.on("end", () => resolve(body));
-              });
-              req.on("error", reject);
-              req.on("timeout", () => { req.destroy(); reject(new Error("Foundry API timeout")); });
-              req.write(postBody);
-              req.end();
-            });
-            // Extract text from Responses API output
-            try {
-              const parsed = JSON.parse(foundryResult);
-              if (parsed.error) {
-                result = `Foundry API error: ${JSON.stringify(parsed.error)}`;
-                log.warn(`AGT sub-agent ${fnName} error: ${result}`);
-              } else {
-                const output = parsed.output || parsed;
-                if (Array.isArray(output)) {
-                  const texts = output
-                    .filter((item: any) => item.type === "message" && item.content)
-                    .flatMap((item: any) => item.content)
-                    .filter((c: any) => c.type === "output_text" || c.type === "text")
-                    .map((c: any) => c.text)
-                    .filter(Boolean);
-                  result = texts.join("\n") || foundryResult;
-                } else {
-                  result = foundryResult;
-                }
-              }
-            } catch {
-              result = foundryResult;
-            }
-            log.info(`AGT sub-agent ${fnName} result: ${result.slice(0, 200)}`);
-          } else if (fnName === "foundry_memory") {
-            // Route through the inference router → Foundry Memory Store API
-            log.info(`AGT sub-agent foundry_memory: ${args.operation} — ${(args.text as string || "").slice(0, 100)}`);
-            const agentName = process.env.SANDBOX_NAME || process.env.HOSTNAME || "default";
-            const store = `memory-${agentName}`;
-            const scope = agentName;
-            const apiVer = "api-version=2025-11-15-preview";
-            const makeItem = (content: string) => ({
-              type: "message", role: "user",
-              content: [{ type: "input_text", text: content }],
-            });
-
-            let memPath: string;
-            let memBody: any;
-            if (args.operation === "update") {
-              memPath = `/memory_stores/${store}:update_memories?${apiVer}`;
-              memBody = { scope, items: [makeItem(args.text as string)], update_delay: 0 };
-            } else {
-              memPath = `/memory_stores/${store}:search_memories?${apiVer}`;
-              memBody = { scope, items: [makeItem(args.text as string)], options: { max_memories: 10 } };
-            }
-
-            const memResult = await new Promise<string>((resolve, reject) => {
-              const postBody = JSON.stringify(memBody);
-              const req = http.request(routerUrl(`${memPath}`), {
-                method: "POST",
-                headers: { "Content-Type": "application/json", "Content-Length": Buffer.byteLength(postBody) },
-                timeout: 30000,
-              }, (res) => {
-                let body = ""; res.on("data", (c: Buffer) => { body += c.toString(); }); res.on("end", () => resolve(body));
-              });
-              req.on("error", reject);
-              req.on("timeout", () => { req.destroy(); reject(new Error("Memory API timeout")); });
-              req.write(postBody);
-              req.end();
-            });
-
-            try {
-              const parsed = JSON.parse(memResult);
-              if (parsed.error) {
-                // Auto-create store on not_found, then retry
-                if ((parsed.error.code === "not_found" || parsed.error.message?.includes("not found")) && args.operation === "update") {
-                  log.info(`Creating memory store '${store}'...`);
-                  const chatModel = process.env.OPENCLAW_MODEL || model;
-                  const createBody = JSON.stringify({
-                    name: store,
-                    description: "AzureClaw sub-agent persistent memory",
-                    definition: { kind: "default", chat_model: chatModel, embedding_model: "text-embedding-3-small",
-                      options: { user_profile_enabled: true, chat_summary_enabled: true } },
-                  });
-                  await new Promise<void>((resolve, reject) => {
-                    const req = http.request(routerUrl(`/memory_stores?${apiVer}`), {
-                      method: "POST",
-                      headers: { "Content-Type": "application/json", "Content-Length": Buffer.byteLength(createBody) },
-                      timeout: 15000,
-                    }, (res) => { let d = ""; res.on("data", (c: Buffer) => { d += c.toString(); }); res.on("end", () => resolve()); });
-                    req.on("error", reject);
-                    req.write(createBody);
-                    req.end();
-                  });
-                  // Retry the original operation
-                  const retryResult = await new Promise<string>((resolve, reject) => {
-                    const postBody = JSON.stringify(memBody);
-                    const req = http.request(routerUrl(`${memPath}`), {
-                      method: "POST",
-                      headers: { "Content-Type": "application/json", "Content-Length": Buffer.byteLength(postBody) },
-                      timeout: 30000,
-                    }, (res) => { let body = ""; res.on("data", (c: Buffer) => { body += c.toString(); }); res.on("end", () => resolve(body)); });
-                    req.on("error", reject);
-                    req.write(postBody);
-                    req.end();
-                  });
-                  result = retryResult;
-                } else {
-                  result = `Memory error: ${JSON.stringify(parsed.error)}`;
-                }
-              } else {
-                if (args.operation === "update") {
-                  result = `Memory updated successfully (id: ${parsed.update_id || parsed.id || "ok"})`;
-                } else {
-                  const memories = parsed.memories || [];
-                  result = memories.length > 0
-                    ? memories.map((m: any) => {
-                        const text = m.memory_item?.content || m.content || m.text || JSON.stringify(m);
-                        return `[${m.score?.toFixed(2) || "?"}] ${text}`;
-                      }).join("\n")
-                    : "No relevant memories found.";
-                }
-              }
-            } catch {
-              result = memResult;
-            }
-            log.info(`AGT sub-agent foundry_memory result: ${result.slice(0, 200)}`);
-          } else if (fnName === "foundry_image_generation") {
-            // Route through the inference router → Azure OpenAI Images API
-            log.info(`AGT sub-agent image_gen: ${(args.prompt as string || "").slice(0, 100)}`);
-            const imgModel = (args.image_model as string) || "gpt-image-1";
-            const imgBody = JSON.stringify({
-              prompt: args.prompt, n: 1,
-              size: args.size || "1024x1024",
-              quality: args.quality || "medium",
-            });
-            const imgResult = await new Promise<string>((resolve, reject) => {
-              const req = http.request(
-                routerUrl(`/openai/deployments/${encodeURIComponent(imgModel)}/images/generations?api-version=2025-04-01-preview`),
-                { method: "POST", headers: { "Content-Type": "application/json", "Content-Length": Buffer.byteLength(imgBody) }, timeout: 120000 },
-                (res) => { let body = ""; res.on("data", (c: Buffer) => { body += c.toString(); }); res.on("end", () => resolve(body)); },
-              );
-              req.on("error", reject);
-              req.on("timeout", () => { req.destroy(); reject(new Error("Image gen timeout")); });
-              req.write(imgBody);
-              req.end();
-            });
-            try {
-              const parsed = JSON.parse(imgResult);
-              const images = parsed?.data || [];
-              const parts: string[] = [];
-              const fs = await import("node:fs");
-              const nodePath = await import("node:path");
-              const os = await import("node:os");
-              const imgDir = nodePath.join(os.tmpdir(), "azureclaw-images");
-              fs.mkdirSync(imgDir, { recursive: true });
-              for (const img of images) {
-                if (img.b64_json) {
-                  const ts = Date.now();
-                  const imgFile = nodePath.join(imgDir, `image-${ts}.png`);
-                  fs.writeFileSync(imgFile, Buffer.from(img.b64_json, "base64"));
-                  parts.push(`📁 Image saved: ${imgFile}`);
-                  if (img.revised_prompt) parts.push(`Revised prompt: ${img.revised_prompt}`);
-                } else if (img.url) {
-                  parts.push(`Image URL: ${img.url}`);
-                }
-              }
-              result = parts.length > 0 ? parts.join("\n") : imgResult;
-            } catch {
-              result = imgResult;
-            }
-            log.info(`AGT sub-agent image_gen complete`);
-          } else if (fnName === "mesh_send") {
-            // Send message to sibling/peer agent via AGT mesh relay (with auto-discover + retry)
-            let toAgent = args.to_agent as string;
-            // OFFLOAD HARDENING: in offload sandboxes, all outbound mesh traffic
-            // is locked to 'parent'. The LLM is explicitly told this in its
-            // system prompt; this is belt-and-braces enforcement.
-            if (process.env.OFFLOAD_REQUEST_ID && toAgent !== "parent") {
-              log.warn(`AGT sub-agent mesh_send: offload mode — rewriting to_agent '${toAgent}' → 'parent' (peer routing disabled in offload)`);
-              toAgent = "parent";
-            }
-            const meshMsg = args.message as string;
-            log.info(`AGT sub-agent mesh_send: to=${toAgent} msg=${(meshMsg || "").slice(0, 100)}`);
-            try {
-              // Single helper handles cache (with TTL) + freshest-match registry
-              // resolution. Retries kept here because sub-agent boots may race
-              // ahead of the peer's registration.
-              //
-              // IMPORTANT: do NOT pass `registryBase: AGT_REGISTRY_URL` — that env
-              // var holds the direct cluster URL (e.g. agentmesh-registry...:8080)
-              // which the egress-guard iptables rules block from the openclaw
-              // container (UID 1000, locked to loopback + DNS). resolveAmidByName's
-              // default routes via the inference-router proxy on 127.0.0.1, which
-              // has its own egress allow-list and reaches the registry correctly.
-              let targetAmid = await resolveAmidByName(toAgent);
-              if (targetAmid) {
-                log.info(`AGT sub-agent mesh_send: resolved AMID for '${toAgent}' (${targetAmid.slice(0, 12)}...)`);
-              }
-
-              for (let attempt = 1; attempt < 8 && !targetAmid; attempt++) {
-                log.info(`AGT sub-agent mesh_send: waiting for '${toAgent}' to register (${attempt}/7)...`);
-                await new Promise(r => setTimeout(r, 2000));
-                targetAmid = await resolveAmidByName(toAgent, { bypassCache: true });
-              }
-
-              if (!targetAmid) {
-                result = `Agent '${toAgent}' not found in registry after retries. It may not be running yet.`;
-              } else if (agtMeshClient) {
-                // Retry send — target may still be uploading prekeys after registration
-                let sendErr: Error | null = null;
-                for (let sendAttempt = 0; sendAttempt < 5; sendAttempt++) {
-                  try {
-                    await agtMeshClient.send(targetAmid, {
-                      type: "task_request",
-                      content: meshMsg,
-                      from_agent: process.env.SANDBOX_NAME || "unknown",
-                      timestamp: new Date().toISOString(),
-                    });
-                    sendErr = null;
-                    break;
-                  } catch (e: any) {
-                    sendErr = e;
-                    if (e.message?.includes("prekey") || e.message?.includes("prekeys")) {
-                      log.info(`AGT sub-agent mesh_send: waiting for prekeys from '${toAgent}' (${sendAttempt + 1}/5)...`);
-                      await new Promise(r => setTimeout(r, 1000));
-                    } else {
-                      break;
-                    }
-                  }
-                }
-                result = sendErr
-                  ? `mesh_send to ${toAgent} failed: ${sendErr.message}`
-                  : `Message sent to ${toAgent} via E2E encrypted mesh relay`;
-              } else {
-                result = `Mesh client not available — cannot send to ${toAgent}`;
-              }
-            } catch (meshErr: any) {
-              result = `mesh_send failed: ${meshErr.message}`;
-            }
-          } else if (fnName === "discover") {
-            // Discover agents in the mesh network
-            const pattern = (args.pattern as string) || "*";
-            log.info(`AGT sub-agent discover: pattern=${pattern}`);
-            try {
-              // Use loopback router proxy — AGT_REGISTRY_URL points at a direct
-              // cluster service that egress-guard blocks for the openclaw container.
-              const registryBase = routerUrl("/agt/registry");
-              const discoverResult = await new Promise<string>((resolve, reject) => {
-                const req = http.get(`${registryBase}/registry/search?capability=${encodeURIComponent(pattern)}`, { timeout: 10000 }, (res) => {
-                  let body = ""; res.on("data", (c: Buffer) => { body += c.toString(); }); res.on("end", () => resolve(body));
-                });
-                req.on("error", reject);
-                req.on("timeout", () => { req.destroy(); reject(new Error("Registry lookup timeout")); });
-              });
-              result = discoverResult;
-            } catch (discErr: any) {
-              result = `discover failed: ${discErr.message}`;
-            }
-          } else if (fnName === "mesh_inbox") {
-            // Check for incoming messages via AGT mesh
-            log.info("AGT sub-agent mesh_inbox check");
-            try {
-              if (agtMeshClient && typeof agtMeshClient.drain === "function") {
-                const messages = await agtMeshClient.drain();
-                result = messages.length > 0
-                  ? JSON.stringify(messages.map((m: any) => ({ from: m.from_agent || m.sender, content: m.content || m.text, timestamp: m.timestamp })))
-                  : "No pending messages";
-              } else {
-                // Fallback: check via AGT inbox (pending messages stored by router)
-                const agtInbox = (globalThis as any).__agtInbox || [];
-                result = agtInbox.length > 0
-                  ? JSON.stringify(agtInbox)
-                  : "No pending messages";
-              }
-            } catch (inboxErr: any) {
-              result = `mesh_inbox failed: ${inboxErr.message}`;
-            }
-          } else {
-            const cmd = String(args.command || args.cmd || "echo 'no command'");
-            log.info(`AGT sub-agent exec: ${sanitizeLog(cmd, 200)}`);
-            let policyAllowed = true;
-            let policyReason = "";
-            try {
-              const policyHttp = await import("node:http");
-              const policyBody = JSON.stringify({ action: `shell:${cmd}`, context: { tool: "exec_command" } });
-              const policyResult = await new Promise<{ allowed: boolean; reason?: string }>((resolve) => {
-                const req = policyHttp.request(routerUrl("/agt/evaluate"), {
-                  method: "POST", timeout: 2000,
-                  headers: { "Content-Type": "application/json", "Content-Length": Buffer.byteLength(policyBody) },
-                }, (res) => {
-                  let data = "";
-                  res.on("data", (c: Buffer) => { data += c.toString(); });
-                  res.on("end", () => { try { resolve(JSON.parse(data)); } catch { resolve({ allowed: true }); } });
-                });
-                req.on("error", () => resolve({ allowed: true }));
-                req.on("timeout", () => { req.destroy(); resolve({ allowed: true }); });
-                req.write(policyBody);
-                req.end();
-              });
-              policyAllowed = policyResult.allowed !== false;
-              policyReason = policyResult.reason || "";
-            } catch { /* router unavailable — allow */ }
-            if (!policyAllowed) {
-              result = `Blocked by policy: ${policyReason || "denied"}`;
-            } else {
-              result = execSync(cmd, { timeout: 15000, encoding: "utf8", maxBuffer: 64 * 1024 }).trim();
-            }
-          }
-        } catch (e: any) {
-          result = e.stderr || e.stdout || e.message || "Command failed";
-        }
-        messages.push({ role: "tool", tool_call_id: tc.id, content: result });
-      }
-      continue;
-    }
-
-    // No tool calls — return the text response
-    return msg.content || "";
-  }
-
-  return "Sub-agent reached maximum tool-calling rounds (25) without a final response.";
+  }, log);
 }
 
 // ── meshSend: auto-chunking send wrapper ─────────────────────────────────────
